+++ packages/lambda/src/functions/helpers/__mocks__/check-if-render-exists.ts
@@ -1,0 +1,5 @@
+import type {checkIfRenderExists as original} from '../check-if-render-exists';
+
+export const checkIfRenderExists: typeof original = () => {
+	return undefined;
+};
