+++ packages/lambda/src/functions/helpers/get-progress.ts
@@ -4,7 +4,6 @@ import type {RenderProgress} from '../../shared/constants';
 import {
 	chunkKey,
 	encodingProgressKey,
-	initalizedMetadataKey,
 	MAX_EPHEMERAL_STORAGE_IN_MB,
 	renderMetadataKey,
 	rendersPrefix,
@@ -12,6 +11,7 @@ import {
 import {DOCS_URL} from '../../shared/docs-url';
 import {calculateChunkTimes} from './calculate-chunk-times';
 import {estimatePriceFromBucket} from './calculate-price-from-bucket';
+import {checkIfRenderExists} from './check-if-render-exists';
 import {getExpectedOutName} from './expected-out-name';
 import {findOutputFileInBucket} from './find-output-file-in-bucket';
 import {formatCostsInfo} from './format-costs-info';
@@ -105,17 +105,12 @@ export const getProgress = async ({
 		expectedBucketOwner,
 	});
 
-	const initializedExists = Boolean(
-		contents.find((c) => {
-			return c.Key?.startsWith(initalizedMetadataKey(renderId));
-		})
+	checkIfRenderExists(
+		contents,
+		renderId,
+		bucketName,
+		getCurrentRegionInFunction()
 	);
-
-	if (!initializedExists) {
-		throw new TypeError(
-			`No render with ID "${renderId}" found in bucket ${bucketName} and region ${region}`
-		);
-	}
 
 	const renderMetadataExists = Boolean(
 		contents.find((c) => c.Key === renderMetadataKey(renderId))
